Vision/2.methodology/requirements.txt
absl-py==1.4.0
abydos==0.5.0
aenum==3.1.15
aiohttp==3.8.3
aiosignal==1.2.0
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
bert-serving-client==1.10.0
bert4keras==0.11.4
blinker==1.4
blis==0.7.8
boto3==1.24.85
botocore==1.27.85
bs4==0.0.1
cachetools==4.2.4
catalogue==2.0.8
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloc==0.2.5
colorama==0.4.3
colored==1.4.4
confection==0.0.2
contourpy==1.0.5
cryptography==38.0.3
cvss==2.5
cycler==0.11.0
cymem==2.0.6
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.6
distro==1.4.0
dnspython==1.16.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
fasteners==0.14.1
filelock==3.12.0
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.3.3
fonttools==4.37.3
frozenlist==1.3.1
fsspec==2023.1.0
future==0.18.2
gast==0.2.2
gdown==4.7.1
gensim==3.8.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
greenlet==1.1.3
gremlinpython==3.7.1
grpcio==1.51.3
h11==0.14.0
h5py==2.10.0
html2markdown==0.1.7
html2text==2020.1.16
httplib2==0.14.0
huggingface-hub==0.14.1
icecream==2.1.3
idna==3.4
importlib-metadata==6.0.0
interchange==2021.0.4
isodate==0.6.1
itsdangerous==2.1.2
javalang==0.13.0
jeepney==0.8.0
jellyfish==0.9.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
JPype1==1.4.1
kaggle==1.5.6
kashgari==2.0.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
kgtools==0.1.11
kiwisolver==1.4.4
LambdaRankNN==0.1.1
langcodes==3.3.0
Lasagne @ https://github.com/Lasagne/Lasagne/archive/master.zip
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Levenshtein==0.20.5
libclang==15.0.6.1
lockfile==0.12.2
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
monotonic==1.5
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.8
neo4j==5.16.0
nest-asyncio==1.5.8
netifaces==0.10.4
networkx==2.6.3
neuralcoref==4.0
nltk==3.7
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
olefile==0.46
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.3.0
pandoc==2.3
pansi==2020.7.3
paramiko==2.7.1
pathos==0.2.9
pathy==0.6.2
pbr==5.11.0
pexpect==4.6.0
Pillow==7.0.0
plumbum==1.8.1
ply==3.11
pox==0.3.1
ppft==1.7.6.5
preshed==3.0.7
protobuf==3.19.6
ptyprocess==0.7.0
py2neo==2021.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
pydantic==1.9.2
PyGithub==1.56
Pygments==2.5.1
PyJWT==2.6.0
pymacaroons==0.13.0
pymongo==4.3.2
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pypandoc==1.11
pyparsing==3.0.9
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.8.1
python-Levenshtein==0.20.5
python-slugify==6.1.2
pytz==2022.2.1
PyVirtualDisplay==3.0
pyxDamerauLevenshtein==1.7.1
pyxdg==0.26
PyYAML==6.0
pyzmq==24.0.1
rapidfuzz==2.11.1
redis==4.3.4
regex==2022.9.13
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.12.0
SecretStorage==2.3.1
sekg==0.10.4.22
selenium==3.14.0
simplejson==3.16.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.41
srsly==2.4.4
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.15.0
tensorflow-addons==0.11.2
tensorflow-estimator==1.15.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
testresources==2.0.1
text-unidecode==1.3
textdistance==4.5.0
Theano==1.0.1
thinc==8.1.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
tqdm==4.64.1
transformers==4.28.1
tree-sitter==0.20.4
trio==0.22.0
trio-websocket==0.9.2
typeguard==2.13.3
typer==0.4.2
typing_extensions==4.5.0
urllib3==1.26.14
validators==0.20.0
wadllib==1.3.3
wasabi==0.10.1
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.15.0
wsproto==1.2.0
xgboost==1.6.2
xlrd==2.0.1
yarl==1.8.1
zipp==3.15.0
